feat: use ic-stable-structures
