+++ src/bitcoin-metaprotocols-canister/Cargo.toml
@@ -9,6 +9,7 @@ crate-type = ["cdylib"]
 [dependencies]
 ic-cdk = "0.18.5"
 ic-cdk-macros = "0.18.5"
+ic-stable-structures = "0.6.4"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 reqwest = { version = "0.11", default-features = false, features = ["json"] }
